Move llm_utils into an llm module
