+++ autogpt/agent/agent_manager.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 from typing import List
 
 from autogpt.config.config import Config
-from autogpt.llm_utils import create_chat_completion
+from autogpt.llm import create_chat_completion
 from autogpt.singleton import Singleton
 from autogpt.types.openai import Message
