+++ autogpt/commands/write_tests.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 import json
 
 from autogpt.commands.command import command
-from autogpt.llm_utils import call_ai_function
+from autogpt.llm import call_ai_function
 
 
 @command(
