+++ autogpt/llm/__init__.py
@@ -1,0 +1,11 @@
+from autogpt.llm.llm_utils import (
+    call_ai_function,
+    create_chat_completion,
+    get_ada_embedding,
+)
+
+__all__ = [
+    "call_ai_function",
+    "create_chat_completion",
+    "get_ada_embedding",
+]
