+++ autogpt/setup.py
@@ -6,7 +6,7 @@ from colorama import Fore, Style
 from autogpt import utils
 from autogpt.config import Config
 from autogpt.config.ai_config import AIConfig
-from autogpt.llm_utils import create_chat_completion
+from autogpt.llm import create_chat_completion
 from autogpt.logs import logger
 
 CFG = Config()
